+++ extra/recorder_enumerator/src/SystemLinqEnumerable.cs
@@ -58,7 +58,7 @@ namespace Dec.RecorderEnumerator
         {
             recorder.Shared().RecordPrivate(input, field_Source, "source");
             recorder.Shared().RecordPrivate(input, field_Comparer, "comparer");
-            recorder.Shared().RecordPrivate(input, field_Set, "selector");
+            recorder.RecordPrivate(input, field_Set, "selector");
             recorder.Shared().RecordPrivate(input, field_Enumerator, "enumerator");
             recorder.RecordPrivate(input, field_State, "state");
             recorder.SharedIfPossible<T>().RecordPrivate(input, field_Current, "current");
@@ -97,7 +97,7 @@ namespace Dec.RecorderEnumerator
             recorder.Shared().RecordPrivate(input, field_Second, "second");
             recorder.Shared().RecordPrivate(input, field_Comparer, "comparer");
             recorder.Shared().RecordPrivate(input, field_Enumerator, "enumerator");
-            recorder.Shared().RecordPrivate(input, field_Set, "selector");
+            recorder.RecordPrivate(input, field_Set, "selector");
             recorder.RecordPrivate(input, field_State, "state");
             recorder.SharedIfPossible<T>().RecordPrivate(input, field_Current, "current");
         }
@@ -135,7 +135,7 @@ namespace Dec.RecorderEnumerator
             recorder.RecordPrivate(input, field_HeadIndex, "headIndex");
             recorder.Shared().RecordPrivate(input, field_Comparer, "comparer");
             recorder.Shared().RecordPrivate(input, field_Enumerator, "enumerator");
-            recorder.Shared().RecordPrivate(input, field_Set, "selector");
+            recorder.RecordPrivate(input, field_Set, "selector");
             recorder.RecordPrivate(input, field_State, "state");
             recorder.SharedIfPossible<T>().RecordPrivate(input, field_Current, "current");
         }