+++ extra/recorder_enumerator/src/SystemLinqEnumerable_Where.cs
@@ -21,7 +21,7 @@ namespace Dec.RecorderEnumerator
         public override void Write(object input, Recorder recorder)
         {
             recorder.Shared().RecordPrivate(input, field_Source, "source");
-            recorder.Shared().RecordPrivate(input, field_Predicate, "predicate");
+            recorder.RecordPrivate(input, field_Predicate, "predicate");
             recorder.Shared().RecordPrivate(input, field_Enumerator, "enumerator");
             recorder.RecordPrivate(input, field_State, "state");
             recorder.SharedIfPossible<T>().RecordPrivate(input, field_Current, "current");
@@ -54,7 +54,7 @@ namespace Dec.RecorderEnumerator
         public override void Write(object input, Recorder recorder)
         {
             recorder.Shared().RecordPrivate(input, field_Source, "source");
-            recorder.Shared().RecordPrivate(input, field_Predicate, "predicate");
+            recorder.RecordPrivate(input, field_Predicate, "predicate");
             recorder.RecordPrivate(input, field_State, "state");
             recorder.SharedIfPossible<T>().RecordPrivate(input, field_Current, "current");
         }
@@ -87,7 +87,7 @@ namespace Dec.RecorderEnumerator
         public override void Write(object input, Recorder recorder)
         {
             recorder.Shared().RecordPrivate(input, field_Source, "source");
-            recorder.Shared().RecordPrivate(input, field_Predicate, "predicate");
+            recorder.RecordPrivate(input, field_Predicate, "predicate");
             recorder.RecordPrivate(input, field_Enumerator, "enumerator");
             recorder.RecordPrivate(input, field_State, "state");
             recorder.SharedIfPossible<T>().RecordPrivate(input, field_Current, "current");
@@ -122,8 +122,8 @@ namespace Dec.RecorderEnumerator
         public override void Write(object input, Recorder recorder)
         {
             recorder.Shared().RecordPrivate(input, field_Source, "source");
-            recorder.Shared().RecordPrivate(input, field_Predicate, "predicate");
-            recorder.Shared().RecordPrivate(input, field_Selector, "selector");
+            recorder.RecordPrivate(input, field_Predicate, "predicate");
+            recorder.RecordPrivate(input, field_Selector, "selector");
             recorder.Shared().RecordPrivate(input, field_Enumerator, "enumerator");
             recorder.RecordPrivate(input, field_State, "state");
             recorder.SharedIfPossible<T>().RecordPrivate(input, field_Current, "current");
@@ -157,8 +157,8 @@ namespace Dec.RecorderEnumerator
         public override void Write(object input, Recorder recorder)
         {
             recorder.Shared().RecordPrivate(input, field_Source, "source");
-            recorder.Shared().RecordPrivate(input, field_Predicate, "predicate");
-            recorder.Shared().RecordPrivate(input, field_Selector, "selector");
+            recorder.RecordPrivate(input, field_Predicate, "predicate");
+            recorder.RecordPrivate(input, field_Selector, "selector");
             recorder.RecordPrivate(input, field_State, "state");
             recorder.SharedIfPossible<T>().RecordPrivate(input, field_Current, "current");
         }
@@ -192,8 +192,8 @@ namespace Dec.RecorderEnumerator
         public override void Write(object input, Recorder recorder)
         {
             recorder.Shared().RecordPrivate(input, field_Source, "source");
-            recorder.Shared().RecordPrivate(input, field_Predicate, "predicate");
-            recorder.Shared().RecordPrivate(input, field_Selector, "selector");
+            recorder.RecordPrivate(input, field_Predicate, "predicate");
+            recorder.RecordPrivate(input, field_Selector, "selector");
             recorder.RecordPrivate(input, field_Enumerator, "enumerator");
             recorder.RecordPrivate(input, field_State, "state");
             recorder.SharedIfPossible<T>().RecordPrivate(input, field_Current, "current");